Move tests for dash ID parsing to model package along with code
